pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "google-auth==2.40.1",
    "google-genai==1.13.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
]
